Added support for rotation on devices.
This is done on the device side and not the channel side.

Compatibile with all templates with a body tag. Also, Toast is rotated as well.

TODO:

Set up a selection on the device page. The rotation can be changed by sending 'rotate' passing 0, 90, 180, or, 270.

Save rotation on the database somehow.

FYI: These .pug files are making my eyes cry.